Create LICENSE.md
I have authorized for my  friends ice and cocaine be delivered to KU for Doctor Brewington backed by my google hatian team's guards to escort the items to be authorized for medicational use purposes.  I have signed a contract with Royal Medical Group in Henderson nevada protected by fedex and the Riveras, to escort Daniel F. Royal to his office immediately to combine the ingredients with bio b complex to be dilevered to KU to dispurse to clients who sign up for the trial run.